backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
bcrypt
python-jose[cryptography]
python-dotenv
